Refactor to Game class
